+++ battlecode22-scaffold/src/AHolidayBot/CombatUtil.java
@@ -63,4 +63,11 @@ public class CombatUtil extends Util{
         return closestCombatUnit;
 	}
 
+    public static boolean enemyArchonLocationGuessIsFalse(MapLocation enemyArchon) throws GameActionException{
+        for (int j = 0; j < archonCount; j++){
+            if (enemyArchon.equals(archonLocations[j])) return true;
+        }
+        return false;
+    }
+
 }